# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is a simple bookstore server with a book inventory, users and\
    \ a shopping cart."
  version: "1.0.0"
  title: "BookHub"
  contact:
    email: "info@bookhub.it"
  license:
    name: "Apache-2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/api"
...
schemes:
- "https"
- "http"
paths:
...
  /books/{bookId}/events:
    get:
      tags:
      - "event"
      - "book"
      summary: "Get events of a book"
      description: "Given a book Id, returns all the events in which was presented"
      operationId: "bookEventsGET"
      produces:
      - "application/json"
      parameters:
      - name: "bookId"
        in: "path"
        description: "Id of the book to get the events"
        required: true
        type: "integer"
        format: "int64"
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0."
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 20 and cannot exceed\
          \ 500."
        required: false
        type: "integer"
      responses:
        200:
          description: "Successful operations"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Event"
        400:
          description: "Invalid request"
        404:
          description: "No events found!"
      x-swagger-router-controller: "Event"
  /books/{bookId}/reviews:
    get:
      tags:
      - "review"
      - "book"
      summary: "Get reviews of a book"
      description: "Given a book Id, returns all the reviews"
      operationId: "bookReviewsGET"
      produces:
      - "application/json"
      parameters:
      - name: "bookId"
        in: "path"
        description: "Id of the book to get the reviews"
        required: true
        type: "integer"
        format: "int64"
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0."
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 20 and cannot exceed\
          \ 500."
        required: false
        type: "integer"
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Review"
        400:
          description: "Invalid request"
        404:
          description: "No reviews found!"
      x-swagger-router-controller: "Review"
  /user/login:
    post:
      tags:
      - "user"
      summary: "Login"
      description: "Login with a form"
      operationId: "userLoginPOST"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "username"
        in: "formData"
        required: true
        type: "string"
      - name: "password"
        in: "formData"
        required: true
        type: "string"
        format: "password"
      responses:
        200:
          description: "Successful login"
        403:
          description: "Wrong username/password"
      x-swagger-router-controller: "User"
  /user/logout:
    post:
      tags:
      - "user"
      summary: "Logout"
      description: "Login with a form"
      operationId: "userLogoutPOST"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Successful logout"
        400:
          description: "Invalid request"
      x-swagger-router-controller: "User"
  /user/register:
    post:
      tags:
      - "user"
      summary: "Register"
      description: "Register into the store"
      operationId: "userRegisterPOST"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "username"
        in: "formData"
        required: true
        type: "string"
      - name: "password"
        in: "formData"
        required: true
        type: "string"
        format: "password"
      - name: "email"
        in: "formData"
        required: true
        type: "string"
      - name: "firstName"
        in: "formData"
        required: true
        type: "string"
      - name: "surname"
        in: "formData"
        required: true
        type: "string"
      - name: "birthDate"
        in: "formData"
        required: true
        type: "string"
        format: "date"
      responses:
        200:
          description: "Successful registration"
        400:
          description: "Invalid Request!"
      x-swagger-router-controller: "User"
  /user/delete:
    delete:
      tags:
      - "user"
      summary: "Delete user"
      description: "Delete an user's account"
      operationId: "userDeletePOST"
      parameters: []
      responses:
        200:
          description: "Successful operation!"
        400:
          description: "Invalid request"
        403:
          description: "Unauthorized"
      x-swagger-router-controller: "User"
...
  /user/update:
    put:
      tags:
      - "user"
      summary: "Update user's data"
      description: "Put request to modify some of the user's data.  One can modify\
        \ single values, so nothing is required.  If no parameter is set a 400 error\
        \ is thrown."
      operationId: "userModifyPUT"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "username"
        in: "formData"
        required: false
        type: "string"
      - name: "password"
        in: "formData"
        required: false
        type: "string"
        format: "password"
      - name: "email"
        in: "formData"
        required: false
        type: "string"
      - name: "firstName"
        in: "formData"
        required: false
        type: "string"
      - name: "surname"
        in: "formData"
        required: false
        type: "string"
      - name: "birthDate"
        in: "formData"
        required: false
        type: "string"
        format: "date"
      responses:
        200:
          description: "Successful operation"
        400:
          description: "Invalid request"
        403:
          description: "Unathorized"
      x-swagger-router-controller: "User"
  /me:
    get:
      tags:
      - "user"
      summary: "Get current user's data"
      description: "Get current logged in user data"
      operationId: "userGetDetailsGET"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid request!"
        403:
          description: "Not Authorized!"
      x-swagger-router-controller: "User"
  /user/{userId}/reviews:
    get:
      tags:
      - "review"
      - "user"
      summary: "Get reviews from a user"
      description: "Given a user Id, returns all the reviews posted by that user"
      operationId: "userReviewsGET"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: "path"
        description: "Id of the user to get the reviews"
        required: true
        type: "integer"
        format: "int64"
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0."
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 20 and cannot exceed\
          \ 500."
        required: false
        type: "integer"
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Review"
        400:
          description: "Invalid request"
        404:
          description: "No reviews found!"
      x-swagger-router-controller: "Review"
...
  /events:
    get:
      tags:
      - "event"
      summary: "Get events filtered"
      operationId: "eventGET"
      produces:
      - "application/json"
      parameters:
      - name: "nameLike"
        in: "query"
        required: false
        type: "string"
      - name: "dateMin"
        in: "query"
        required: false
        type: "string"
        format: "date-time"
      - name: "dateMax"
        in: "query"
        required: false
        type: "string"
        format: "date-time"
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0."
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 20 and cannot exceed\
          \ 500."
        required: false
        type: "integer"
      responses:
        200:
          description: "Successful operations"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Event"
        400:
          description: "Invalid request"
        404:
          description: "No events found!"
      x-swagger-router-controller: "Event"
  /events/{eventId}:
    get:
      tags:
      - "event"
      summary: "Get an event in which a book is presented"
      operationId: "eventIdGET"
      produces:
      - "application/json"
      parameters:
      - name: "eventId"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Event"
        404:
          description: "Event not found!"
      x-swagger-router-controller: "Event"
  /orders:
    get:
      tags:
      - "order"
      summary: "Current user's orders"
      description: "Collection of orders placed by the current user"
      operationId: "ordersGET"
      produces:
      - "application/json"
      parameters:
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0."
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 20 and cannot exceed\
          \ 500."
        required: false
        type: "integer"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Order"
        400:
          description: "Invalid request"
        403:
          description: "Not authorized!"
      x-swagger-router-controller: "Order"
  /orders/create:
    post:
      tags:
      - "order"
      summary: "Place a new order"
      description: "Place a new order from the cart"
      operationId: "orderPlacePOST"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "addressStreetLine1"
        in: "formData"
        required: true
        type: "string"
      - name: "addressStreetLine2"
        in: "formData"
        required: false
        type: "string"
      - name: "city"
        in: "formData"
        required: true
        type: "string"
      - name: "zip_code"
        in: "formData"
        required: true
        type: "string"
      - name: "province"
        in: "formData"
        required: true
        type: "string"
      - name: "country"
        in: "formData"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful placement"
        400:
          description: "Bad Request"
        403:
          description: "Unauthorized"
      x-swagger-router-controller: "Order"
  /orders/{orderId}/address:
    get:
      tags:
      - "order"
      summary: "Get shipment address of a order"
      description: "Given an order id, returns the shipment address set to that order,\
        \ if user is authorized to see it"
      operationId: "orderAddressGET"
      produces:
      - "application/json"
      parameters:
      - name: "orderId"
        in: "path"
        description: "The id of the order. Must be associated to the current logged\
          \ in user"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Address"
        403:
          description: "Not authorized!"
        404:
          description: "Review not found!"
      x-swagger-router-controller: "Order"
  /reviews/{reviewId}:
    get:
      tags:
      - "review"
      summary: "Get a review by Id"
      operationId: "reviewIdGET"
      produces:
      - "application/json"
      parameters:
      - name: "reviewId"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Review"
        404:
          description: "Review not found!"
      x-swagger-router-controller: "Review"
definitions:
  User:
    type: "object"
    required:
    - "birth_date"
    - "email"
    - "first_name"
    - "surname"
    - "user_id"
    - "username"
    properties:
      user_id:
        type: "integer"
        format: "int64"
        example: 1
      username:
        type: "string"
        example: "malta895"
      first_name:
        type: "string"
        example: "Mario"
      surname:
        type: "string"
        example: "Rossi"
      email:
        type: "string"
        format: "email"
      birth_date:
        type: "string"
        format: "date"
      address:
        $ref: "#/definitions/Address"
    title: "User"
    description: "A user of the application."
    example:
      address:
        country: "Italy"
        province: "CO"
        city: "Como"
        address_id: 1
        street_line2: "11"
        street_line1: "Via Valleggio"
        zip_code: "22100"
      user_id: 1
      surname: "Rossi"
      birth_date: "2000-01-23"
      first_name: "Mario"
      email: "email"
      username: "malta895"
...
  Book:
    type: "object"
    required:
    - "book_id"
    - "price"
    - "publisher"
    - "title"
    properties:
      book_id:
        type: "integer"
        format: "int64"
        example: 1
      isbn:
        type: "string"
        example: "123456679"
      title:
        type: "string"
        example: "Il deserto dei tartari"
      authors:
        type: "array"
        items:
          $ref: "#/definitions/Author"
      price:
        $ref: "#/definitions/Amount"
      picture_path:
        type: "string"
        example: "pic1.jpg"
        description: "Relative path to a picture representing the book\n"
      status:
        type: "string"
        example: "Available"
        description: "Book availability in the inventory"
        enum:
        - "Available"
        - "Out of stock"
      publisher:
        $ref: "#/definitions/Publisher"
      average_rating:
        type: "number"
        format: "double"
        description: "Average rating of the book, calculated from the reviews"
      genre:
        $ref: "#/definitions/Genre"
      theme:
        $ref: "#/definitions/Theme"
    title: "Book"
    description: "A book for sale in the store"
    example:
      picture_path: "pic1.jpg"
      price:
        currency: "EUR"
        value: 11
      isbn: "123456679"
      genre:
        name: "Horror"
        description: "Very scary"
        genre_id: 1
      publisher:
        publisher_id: 1
        hq_location:
          country: "Italy"
          province: "CO"
          city: "Como"
          address_id: 1
          street_line2: "11"
          street_line1: "Via Valleggio"
          zip_code: "22100"
        name: "Zanichelli"
        description: "description"
      average_rating: 0.80082819046101150206595775671303272247314453125
      theme:
        name: "name"
        description: "description"
      book_id: 1
      title: "Il deserto dei tartari"
      authors:
      - surname: "Rossi"
        name: "Mario"
        biography: "biography"
        picture: "picture"
      - surname: "Rossi"
        name: "Mario"
        biography: "biography"
        picture: "picture"
      status: "Available"
  Amount:
    type: "object"
    required:
    - "currency"
    - "value"
    properties:
      value:
        type: "number"
        format: "double"
        minimum: 0.01
        maximum: 1000000000
      currency:
        $ref: "#/definitions/Currency"
    description: "Price of an item"
    example:
      currency: "EUR"
      value: 11
  Currency:
    type: "string"
    pattern: "^[A-Z]{3,3}$"
    description: "Currency international code\n"
    example: "EUR"
  Author:
    type: "object"
    required:
    - "name"
    - "surname"
    properties:
      name:
        type: "string"
        example: "Mario"
        description: "First name of the author\n"
      surname:
        type: "string"
        example: "Rossi"
        description: "Surname of the author\n"
      biography:
        type: "string"
        description: "Short biography of the author\n"
      picture:
        type: "string"
        description: "A portrait of the author\n"
    title: "Author"
    description: "An author that has written or co-written some books in the store"
    example:
      surname: "Rossi"
      name: "Mario"
      biography: "biography"
      picture: "picture"
  Address:
    type: "object"
    required:
    - "city"
    - "country"
    - "street_line1"
    - "zip_code"
    properties:
      address_id:
        type: "integer"
        format: "int64"
        example: 1
      street_line1:
        type: "string"
        example: "Via Valleggio"
      street_line2:
        type: "string"
        example: "11"
      city:
        type: "string"
        example: "Como"
      zip_code:
        type: "string"
        example: "22100"
      province:
        type: "string"
        example: "CO"
      country:
        type: "string"
        example: "Italy"
    title: "Address"
    description: "A complete address template, usable in many contexts of the web\
      \ app"
    example:
      country: "Italy"
      province: "CO"
      city: "Como"
      address_id: 1
      street_line2: "11"
      street_line1: "Via Valleggio"
      zip_code: "22100"
  Event:
    type: "object"
    required:
    - "date"
    - "event_id"
    - "location"
    - "name"
    - "presented_book"
    properties:
      event_id:
        type: "integer"
        format: "int64"
        example: 1
        description: "Unique identifier of the event"
      location:
        $ref: "#/definitions/Address"
      presented_book:
        $ref: "#/definitions/Book"
      name:
        type: "string"
      date:
        type: "string"
        format: "date-time"
    title: "Event"
    description: "An event where some books are presented"
    example:
      date: "2000-01-23T04:56:07.000+00:00"
      event_id: 1
      name: "name"
      location:
        country: "Italy"
        province: "CO"
        city: "Como"
        address_id: 1
        street_line2: "11"
        street_line1: "Via Valleggio"
        zip_code: "22100"
      presented_book:
        picture_path: "pic1.jpg"
        price:
          currency: "EUR"
          value: 11
        isbn: "123456679"
        genre:
          name: "Horror"
          description: "Very scary"
          genre_id: 1
        publisher:
          publisher_id: 1
          hq_location:
            country: "Italy"
            province: "CO"
            city: "Como"
            address_id: 1
            street_line2: "11"
            street_line1: "Via Valleggio"
            zip_code: "22100"
          name: "Zanichelli"
          description: "description"
        average_rating: 0.80082819046101150206595775671303272247314453125
        theme:
          name: "name"
          description: "description"
        book_id: 1
        title: "Il deserto dei tartari"
        authors:
        - surname: "Rossi"
          name: "Mario"
          biography: "biography"
          picture: "picture"
        - surname: "Rossi"
          name: "Mario"
          biography: "biography"
          picture: "picture"
        status: "Available"
  Publisher:
    type: "object"
    required:
    - "name"
    - "publisher_id"
    properties:
      publisher_id:
        type: "integer"
        format: "int64"
        example: 1
      hq_location:
        $ref: "#/definitions/Address"
      name:
        type: "string"
        example: "Zanichelli"
        description: "Name of the editor/publisher"
      description:
        type: "string"
        description: "Some information on the publisher"
    title: "Publisher"
    description: "Who edits/publishes a book"
    example:
      publisher_id: 1
      hq_location:
        country: "Italy"
        province: "CO"
        city: "Como"
        address_id: 1
        street_line2: "11"
        street_line1: "Via Valleggio"
        zip_code: "22100"
      name: "Zanichelli"
      description: "description"
  Genre:
    type: "object"
    required:
    - "genre_id"
    - "name"
    properties:
      genre_id:
        type: "integer"
        format: "int64"
        example: 1
      name:
        type: "string"
        example: "Horror"
      description:
        type: "string"
        example: "Very scary"
    title: "Genre"
    description: "The genre of the book"
    example:
      name: "Horror"
      description: "Very scary"
      genre_id: 1
  Theme:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
      description:
        type: "string"
    title: "Theme"
    description: "The theme of the book"
    example:
      name: "name"
      description: "description"
  Review:
    type: "object"
    required:
    - "book"
    - "rating"
    - "review_id"
    - "user"
    properties:
      review_id:
        type: "integer"
        format: "int64"
      user:
        type: "integer"
        format: "int64"
        description: "User id of who wrote the review"
      title:
        type: "string"
        description: "Title of the review"
      text:
        type: "string"
        description: "Content of the review"
      rating:
        type: "integer"
        format: "int64"
        description: "Rating of the review, expressed in 1 to 5 stars"
        minimum: 1
        maximum: 5
      book:
        type: "integer"
        format: "int64"
        description: "Id of the reviewed book"
    title: "Review"
    description: "A review of a book, with rating and optional title and text"
    example:
      review_id: 0
      book: 5
      rating: 1
      text: "text"
      title: "title"
      user: 6
  Order:
    type: "object"
    required:
    - "items"
    - "order_id"
    - "shipment_address"
    - "total_amount"
    - "user_id"
    properties:
      order_id:
        type: "integer"
        format: "int64"
        description: "Id of the order"
      user_id:
        type: "integer"
        format: "int64"
        description: "UserId of who placed the order"
      items:
        type: "array"
        items:
          $ref: "#/definitions/Book"
      total_amount:
        $ref: "#/definitions/Amount"
      shipment_address:
        $ref: "#/definitions/Address"
      order_date:
        type: "string"
        format: "date-time"
    example:
      order_date: "2000-01-23T04:56:07.000+00:00"
      user_id: 6
      total_amount:
        currency: "EUR"
        value: 11
      shipment_address:
        country: "Italy"
        province: "CO"
        city: "Como"
        address_id: 1
        street_line2: "11"
        street_line1: "Via Valleggio"
        zip_code: "22100"
      order_id: 0
      items:
      - picture_path: "pic1.jpg"
        price:
          currency: "EUR"
          value: 11
        isbn: "123456679"
        genre:
          name: "Horror"
          description: "Very scary"
          genre_id: 1
        publisher:
          publisher_id: 1
          hq_location:
            country: "Italy"
            province: "CO"
            city: "Como"
            address_id: 1
            street_line2: "11"
            street_line1: "Via Valleggio"
            zip_code: "22100"
          name: "Zanichelli"
          description: "description"
        average_rating: 0.80082819046101150206595775671303272247314453125
        theme:
          name: "name"
          description: "description"
        book_id: 1
        title: "Il deserto dei tartari"
        authors:
        - surname: "Rossi"
          name: "Mario"
          biography: "biography"
          picture: "picture"
        - surname: "Rossi"
          name: "Mario"
          biography: "biography"
          picture: "picture"
        status: "Available"
      - picture_path: "pic1.jpg"
        price:
          currency: "EUR"
          value: 11
        isbn: "123456679"
        genre:
          name: "Horror"
          description: "Very scary"
          genre_id: 1
        publisher:
          publisher_id: 1
          hq_location:
            country: "Italy"
            province: "CO"
            city: "Como"
            address_id: 1
            street_line2: "11"
            street_line1: "Via Valleggio"
            zip_code: "22100"
          name: "Zanichelli"
          description: "description"
        average_rating: 0.80082819046101150206595775671303272247314453125
        theme:
          name: "name"
          description: "description"
        book_id: 1
        title: "Il deserto dei tartari"
        authors:
        - surname: "Rossi"
          name: "Mario"
          biography: "biography"
          picture: "picture"
        - surname: "Rossi"
          name: "Mario"
          biography: "biography"
          picture: "picture"
        status: "Available"
...